requirements.txt
django-cors-headers==3.13.0
django-filter==22.1
Django==3.2.14
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
